CVNAKE GAME (using OpenCV on Python)
It has a spontaneous changing score table consisting of users played from the same computer with the local database connection.

The database contains username, password, maximum score and total playing time.

In case of hitting walls or eating herself, the snake loses 3 points and the length of the snake is shortened by 3 units. User has 3 chance and hearts decrease as life decreases